af-packet: move into source/af-packet module
